refactor to pull out file menu into own component
