+++ UI/src/components/CodeEditor/CodeEditor.tsx
@@ -20,20 +20,12 @@ import AddFileModal from "../AddFileModal/AddFileModal";
 import DockerLogsViewer from "../DockerLogsViewer/DockerLogsViewer";
 import FileEditor from "../FileEditor/FileEditor";
 import UploadStatusModal from "../UploadStatusModal/UploadStatusModal";
+import CodeEditorMenu from "./CodeEditorMenu";
 
 const CodeEditor = () => {
-  const [collapsed, setCollapsed] = useState(false);
-  const [showAddFileModal, setShowAddFileModal] = useState(false);
   const [showUploadStatusModal, setShowUploadStatusModal] = useState(false);
   const [showDockerLogsModal, setShowDockerLogsModal] = useState(false);
-  const [messageApi, contextHolder] = message.useMessage();
-  const {
-    getFileByPath,
-    setSelectedFile,
-    selectedFile,
-    files,
-    selectedBackend,
-  } = appStore();
+  const { selectedFile, files, selectedBackend } = appStore();
 
   const {
     redeployProject,
@@ -45,23 +37,6 @@ const CodeEditor = () => {
   } = useDockerService();
   const [hasUploaded, setHasUploaded] = useState(Boolean(link));
   const appSlug = appSlugFromURL(link);
-  const menuItems = buildMenuItemsFromFiles(files);
-
-  const handleMenuItemClick: MenuProps["onClick"] = (e) => {
-    if (!e) return;
-    const filePath = String(e.key).toLowerCase();
-    console.log(filePath);
-    const file = getFileByPath(filePath);
-    if (!file) {
-      messageApi.error(`Something went wrong accessing: ${e.key}`);
-      return;
-    }
-    setSelectedFile(file);
-  };
-
-  const handleNewFileButton = () => {
-    setShowAddFileModal(true);
-  };
 
   const handleRunButton = () => {
     const asDir = buildDirectoryTree(files);
@@ -78,7 +53,7 @@ const CodeEditor = () => {
       setShowUploadStatusModal(true);
       return;
     }
-    messageApi.error("error getting app slug on redeploy");
+    message.error("error getting app slug on redeploy");
   };
 
   const items1: MenuProps["items"] = [
@@ -87,7 +62,6 @@ const CodeEditor = () => {
   ];
   return (
     <Layout style={{ minHeight: "100vh", minWidth: "100vh" }}>
-      {contextHolder}
       <Header style={{ display: "flex", alignItems: "center" }}>
         <div style={{ marginLeft: "13.5vw" }} />
         <Menu
@@ -130,52 +104,7 @@ const CodeEditor = () => {
         />
       </Header>
       <Layout>
-        <Sider
-          collapsible
-          collapsed={collapsed}
-          onCollapse={(value) => setCollapsed(value)}
-        >
-          <div
-            style={{
-              display: "flex",
-              flexDirection: "row",
-              justifyContent: "space-around",
-              marginBottom: "10%",
-            }}
-          >
-            <Button
-              style={{ width: "40%", backgroundColor: "var(--gray)" }}
-              icon={<UploadOutlined />}
-              type="primary"
-            />
-            <Button
-              style={{ width: "40%", backgroundColor: "var(--dun)" }}
-              icon={<DownloadOutlined style={{ color: "black" }} />}
-              type="primary"
-            />
-          </div>
-          <Button
-            style={{
-              width: "80%",
-              marginLeft: "10%",
-              marginBottom: "10%",
-              backgroundColor: "var(--anti-flash-white)",
-            }}
-            onClick={handleNewFileButton}
-            icon={<FileAddOutlined />}
-          />
-          <Menu
-            theme="dark"
-            defaultSelectedKeys={["1"]}
-            mode="inline"
-            items={menuItems}
-            onClick={handleMenuItemClick}
-          />
-          <AddFileModal
-            showModal={showAddFileModal}
-            setShowModal={setShowAddFileModal}
-          />
-        </Sider>
+        <CodeEditorMenu />
         <Layout>
           <Content style={{ width: "100%", height: "100%" }}>
             {selectedFile ? <FileEditor /> : <></>}
